move sampling to cycle profiler.
