+++ src/execution/operator/aggregate/physical_simple_aggregate.cpp
@@ -66,7 +66,7 @@ public:
 
 class SimpleAggregateLocalState : public LocalSinkState {
 public:
-	explicit SimpleAggregateLocalState(vector<unique_ptr<Expression>> &aggregates) : state(aggregates) {
+	explicit SimpleAggregateLocalState(vector<unique_ptr<Expression>> &aggregates) : state(aggregates), function_counter() {
 		vector<LogicalType> payload_types;
 		for (auto &aggregate : aggregates) {
 			D_ASSERT(aggregate->GetExpressionClass() == ExpressionClass::BOUND_AGGREGATE);
@@ -142,11 +142,11 @@ void PhysicalSimpleAggregate::Sink(ExecutionContext &context, GlobalOperatorStat
 			}
 		}
 
-		sink.function_counter.Start();
+		sink.function_counter.TakeSample();
 		aggregate.function.simple_update(payload_cnt == 0 ? nullptr : &payload_chunk.data[payload_idx],
 		                                 aggregate.bind_info.get(), payload_cnt, sink.state.aggregates[aggr_idx].get(),
 		                                 payload_chunk.size());
-		sink.function_counter.End();
+		sink.function_counter.ProcessSample();
 		payload_idx += payload_cnt;
 	}
 }
