+++ src/include/duckdb/common/cycle_counter.hpp
@@ -10,32 +10,72 @@
 
 #include "duckdb/common/helper.hpp"
 #include "duckdb/common/chrono.hpp"
+#include "duckdb/common/random_engine.hpp"
 
 namespace duckdb {
 
-//! The cycle counter can be used to measure elapsed cycles
+//! The cycle counter can be used to measure elapsed cycles for a function, expression and ...
+//! Optimized by sampling mechanism. Once per 100 times.
+//! //Todo Can be optimized further by calling RDTSC once per sample
 class CycleCounter {
 public:
-	//! Starts the timer
-	void Start() {
-		finished = false;
-		start = Tick();
+	CycleCounter() : random(0) {
 	}
-	//! Finishes timing
-	void End() {
-		end = Tick();
-		finished = true;
-	}
+	// Next_sample determines if a sample needs to be taken, if so start the profiler
+    void TakeSample() {
+        if (current_count >= next_sample) {
+            tmp = Tick();
+        }
+    }
 
-	uint64_t Elapsed() const {
-		return end - start;
-	}
+    // Process the sample
+    void ProcessSample(int chunk_size) {
+        if (current_count >= next_sample) {
+            time += Tick() - tmp;
+        }
+        if (current_count >= next_sample) {
+            next_sample = 50 + random.NextRandomInteger() % 100;
+            ++sample_count;
+            sample_tuples_count += chunk_size;
+            current_count = 0;
+        } else {
+            ++current_count;
+        }
+        tuples_count += chunk_size;
+    }
 
-private:
+    // Process the sample
+    void ProcessSample() {
+        if (current_count >= next_sample) {
+            time += Tick() - tmp;
+        }
+        if (current_count >= next_sample) {
+            next_sample = 50 + random.NextRandomInteger() % 100;
+            ++sample_count;
+            current_count = 0;
+        } else {
+            ++current_count;
+        }
+    }
+
+public:
 	uint64_t Tick() const;
-	uint64_t start;
-	uint64_t end;
-	bool finished = false;
+	// current number on RDT register
+	uint64_t tmp;
+	// Elapsed cycles
+	uint64_t time = 0;
+    //! Count the number of time the executor called since last sampling
+    uint64_t current_count = 0;
+    //! Show the next sample
+    uint64_t next_sample = 0;
+    //! Count the number of samples
+    uint64_t sample_count = 0;
+    //! Count the number of tuples sampled
+    uint64_t sample_tuples_count = 0;
+    //! Count the number of ALL tuples
+    uint64_t tuples_count = 0;
+    //! the random number generator used for sampling
+    RandomEngine random;
 };
 
 } // namespace duckdb
