+++ src/include/duckdb/execution/expression_executor_state.hpp
@@ -29,7 +29,6 @@ struct ExpressionState {
 	vector<LogicalType> types;
 	DataChunk intermediate_chunk;
 	string name;
-	double time;
 	CycleCounter profiler;
 
 public:
@@ -43,46 +42,6 @@ struct ExpressionExecutorState {
 	ExpressionExecutor *executor;
 	CycleCounter profiler;
 	string name;
-	double time;
-	//! Count the number of time the executor called
-	uint64_t total_count = 0;
-	//! Count the number of time the executor called since last sampling
-	uint64_t current_count = 0;
-	//! Show the next sample
-	uint64_t next_sample = 0;
-	//! Count the number of samples
-	uint64_t sample_count = 0;
-	//! Count the number of tuples in all samples
-	uint64_t sample_tuples_count = 0;
-	//! Count the number of tuples processed by this executor
-	uint64_t tuples_count = 0;
-	//! the random number generator used for sampling
-	RandomEngine random;
-
-	// Next_sample determines if a sample needs to be take, if so start the profiler
-	void TakeSample() {
-		if (current_count >= next_sample) {
-			profiler.Start();
-		}
-	}
-
-	// Process the sample
-	void ProcessSample(int chunk_size) {
-		if (current_count >= next_sample) {
-			profiler.End();
-			time += profiler.Elapsed();
-		}
-		if (current_count >= next_sample) {
-			next_sample = 50 + random.NextRandomInteger() % 100;
-			++sample_count;
-			sample_tuples_count += chunk_size;
-			current_count = 0;
-		} else {
-			++current_count;
-		}
-		++total_count;
-		tuples_count += chunk_size;
-	}
 };
 
 } // namespace duckdb
